Move Players datatype into Data_Models pre-split
