+++ aesopstables.py
@@ -1,6 +1,6 @@
 from aesops import app, db
 from pairing.match import Match
-from pairing.player import Player
+from data_models.players import Player
 from data_models.tournaments import Tournament
 from data_models.users import User
