+++ documentation/pairing_strategy.md
